Dev work veiw styling
Working on lazy-loading.  Theere is an issue where in within the panelizer template - when a regions' content is called it auto wraps the content in a div [ 'block-region-(region-name)' ].  This is cauing conflict since we want a floating title in the splash-right region, and lazy load in content as you scroll. The extra wrapper div is causing an issue with the position styling.